pyproject.toml
[project]
name = "medallion"
version = "0.1.0"
description = ""
authors = [
    {name = "mcklmo",email = "m.h98@gx.net"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pydantic (>=2.13.4,<3.0.0)",
    "google-auth (>=2.52.0,<3.0.0)",
    "pendulum (>=3.2.0,<4.0.0)",
    "google-cloud-storage (>=3.10.1,<4.0.0)"
]

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
